Adds chat and appointment manger buttons to Provider Dash. Also adds basic structure of dash. Moves MedicationList component to shared components as it will be reused in provider dash
